requirements.txt
ipykernel
selenium
python-dotenv
regex
transformers
matplotlib
scikit-learn
pandas
transformers[torch]